[BE] refactor(#114) : 피드백 반영
* StudyInfo : joinCode생성 메소드를 도메인 영역으로 이동
* StudyInfoService : 복잡한 로직들 함수화 리펙토링
