+++ backend/src/main/java/com/example/backend/domain/define/study/info/StudyInfo.java
@@ -12,13 +12,18 @@ import lombok.NoArgsConstructor;
 import org.hibernate.annotations.ColumnDefault;
 import org.hibernate.annotations.DynamicInsert;
 
+import java.security.SecureRandom;
 import java.time.LocalDate;
+import java.util.Random;
 
 @Getter
 @DynamicInsert
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity(name = "STUDY_INFO")
 public class StudyInfo extends BaseEntity {
+    private static final String CHARACTERS = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
+    public static final int JOIN_CODE_LENGTH = 10;
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "STUDY_INFO_ID")
@@ -80,7 +85,7 @@ public class StudyInfo extends BaseEntity {
         this.endDate = endDate;
         this.info = info;
         this.status = status;
-        this.joinCode = joinCode;
+        this.joinCode = generateRandomString(JOIN_CODE_LENGTH);
         this.maximumMember = maximumMember;
         this.currentMember = currentMember;
         this.lastCommitDay = lastCommitDay;
@@ -89,4 +94,14 @@ public class StudyInfo extends BaseEntity {
         this.repositoryInfo = repositoryInfo;
         this.periodType = periodType;
     }
+    private String generateRandomString(int length) {
+        Random random = new SecureRandom();
+        StringBuilder sb = new StringBuilder(length);
+        for (int i = 0; i < length; i++) {
+            sb.append(CHARACTERS.charAt(random.nextInt(CHARACTERS.length())));
+        }
+        return sb.toString();
+    }
+
+
 }
